Delete post
Should post need to be deleted, it will need an endpoint to do that; that's the essence of this commit. A DELETE enpoint has been added and given a valid id, will delete the specified id. In cases where the id is incorrect, it throws an error showing that the id is non-existent.